// class02/car-dealership/src/app/types/car.interface.ts
export interface Car {
  id: string; // e.g. '1', '2', etc.
  description: string; // e.g. 'Како нова, баба ја возела', etc.
  price: number; // e.g. 100000, 200000, etc.
  brand: string; // e.g. 'BMW', 'Mercedes', 'Audi', etc. (or use an enum)
  model: string; // e.g. 'X5', 'X7', 'S8', etc.
}